Add files via upload
The purpose of this project was to analyze decarbonization efforts in the context of achieving a net-zero carbon economy. By leveraging a simulated decarbonization dataset, we explored key factors such as energy consumption, carbon emissions, and carbon cost per ton of CO2. The goal was to understand how these variables relate to each other and how they can be used to predict carbon pricing, which can drive further decision-making in policy and industry to reduce carbon emissions leading to a net zero carbon economy.